backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
pillow==9.5.0
transformers
torch
gunicorn
torchvision
scikit-image